Add files via upload
Q. An online shopping store is providing discounts on the items due to the Eid. If the cost of items is less than 2000 it will give a discount up to 5%. If the cost of shopping is 2000 to 4000, a 10% discount will be applied. If the cost of shopping is 4000 to 6000, a 20% discount will be applied. If it's more than 6000 then a 35% discount will be applied to the cost of shopping. Print the actual amount,saved amount and the amount after discount. TheMinimum amount eligible for a discount is 500